test(ai): replace exported wrapTracer unit tests with E2E regression in index.spec.js
